requirements.txt
streamlit
pandas
langchain
langchain-community
langchain-google-genai
sqlalchemy
python-dotenv
tabulate